bin/fortune: Homebrew fortune-printing program
We use this because it saves installing `/usr/bin/fortune` on systems.
(It probably wouldn't even be considered appropriate for many hosts I
admin.)

It's far from efficient, but when we're looking at a 1500-line fortune
database, it makes no difference on even semi-modern machines.

This allows us to get rid of the whole `.dat`-file generation thing.