vimrc: Stop using swap files, move gX mapping
Swap files are annoying when using git. Earlier I had moved them to
a location relative to home, so it wouldnt interfere with my VCS.
That configuration doesn't work on this Ubuntu system I'm using.
Instead of bothering to debug it, I'm just going to disable it
completely, since I save my files often enough for this to not be
risky.